Merge kroah.com:/home/linux/linux/BK/bleeding-2.5 into kroah.com:/home/linux/linux/BK/gregkh-2.5
2002/11/27 10:56:01-08:00 stuartm
[PATCH] WhiteHEAT update

1-fix-lowlat:

QA found that running all four ports at 460800 would drop data. I
traced it to data being dropped in the read callback because the flip
buffers were full. Turning on the low latency flag fixed things.

2-fix-taint

A side-effect of turning on low latency; the interrupt context from
the callback is now passed through to the tty layer, passing it on to
calls back into usb-serial.c. Which causes deadlocks when trying to
re-acquire the per-port semaphore. We've already talked about this.
This patch is my work-around for the usb-serial.c brokenness.
Basically, implemement a buffering scheme, and schedule a software
interrupt to handle the data handoff to the tty layer sometime later.
urb_pool_size defaults to 8, but is a module parameter and can be
modified at runtime.
The buffering is needed so that the driver can run while data is
waiting to be processed, but I could have used the tty layer
scheduling instead of doing my own by turning off low latency.
However, I looked at the tty layer and it seems to me that there's
nothing preventing a really fast device from flipping one buffer,
flipping the next, and flipping back to the still full buffer from
before (actually, the flip just gets scheduled for later), so my
driver needs to be able to hold onto buffered data and schedule them
for processing later anyway. So, might as well leave low_latency on.


diff -Naur linux-2.5.49-0-virgin/drivers/usb/serial/whiteheat.c linux-2.5.49-1-fix-
lowlat/drivers/usb/serial/whiteheat.c

2002/11/27 10:27:41-08:00 stuartm
[PATCH] usb-serial.c disconnect race

Attached is a patch that changes the 2.5.x disconnect to be similar to
2.4.x disconnect. This doesn't fix the race, but does shrink the
window such that I've never seen it trigger, even under testing
designed to do that.

There doesn't seem to be a good way to fix the race. The fix should be
to have _disconnect force any sleeping semaphore holders to run to
completion between the end of the loop in the patch below and the spot
where the underlying memory is freed, but I don't see a way to do
that.


diff -Naur linux-2.5.49-2-fix-taint/drivers/usb/serial/usb-serial.c linux-2.5.49-3-fix-drvdata/drivers/usb/serial/usb-
serial.c

2002/11/27 09:54:32-08:00 jtyner
[PATCH] [patch] speed/clean up vicam_decode_color

This patch cleans up the vicam_decode_color function by removing
unused/useless variables and combining the two "x" loops inside the
y loop into one. It also reduces the number of times that the "x"
loop occurs from 512 to 320 which should provide a decent speed
increase. It also fixes a bug in the y loop that wrote beyond its bound.

2002/11/27 09:46:14-08:00 nemosoft
[PATCH] [PATCH] PWC 8.9

After a little absence, here's a patch to bring the Philips Webcam driver up
to version 8.9 (skipping 8.8 which has been available as a download on my
website for a while).

This patch is against 2.5.49, and includes some of the following:
* New USB IDs for Logitech and Visionite webcams.
* Better URB link/unlink sequence when opening/closing device
  and switching resolutions.
* Adding probe for CCD/CMOS sensor type.
* Removed remnants of YUV420 palette stuff.

Also updated the description in 'Kconfig'.

2002/11/27 09:40:14-08:00 mark
[PATCH] USB ov511 driver: Update to version 1.63

2002/11/27 09:06:12-08:00 ganesh
[PATCH] added support for insmod options to specify vendor/product id. this

2002/11/26 15:33:15-08:00 greg
[PATCH] USB serial: split the generic functions out into their own file


2002/11/26 14:27:06-08:00 randy.dunlap
[PATCH] tiglusb timeouts

It addresses the timeout parameter in the tiglusb driver.

1.  timeout could be 0, causing a divide-by-zero.
The patch prevents this.

2.  The timeout value to usb_bulk_msg() could be rounded
down to cause a divide-by-zero if timeout was < 10, e.g. 9,
in:
	result = usb_bulk_msg (s->dev, pipe, buffer, bytes_to_read,
			       &bytes_read, HZ / (timeout / 10));
9 / 10 == 0 => divide-by-zero !!

3.  The timeout value above doesn't do very well on converting
timeout to tenths of seconds.  Even for the default timeout
value of 15 (1.5 seconds), it becomes:
	HZ / (15 / 10) == HZ / 1 == HZ, or 1 second.
The patch corrects this formula to use:
	(HZ * 10) / timeout

2002/11/26 14:16:49-08:00 greg
[PATCH] USB serial: move the ezusb functions into their own file.


2002/11/25 16:53:04-08:00 jtyner
[PATCH] USB: [patch] fix vicam disconnect/locking

Here is a patch that fixes the disconnect handling and locking for the
vicam driver. It does the following.

1.)	Change the parameters of send_control_msg to take a struct
	vicam_camera instead of struct usb_device to allow for locking
	of the device. Note that __send_control_msg does not lock the
	camera. send_control_msg locks the camera before calling
	__send_control_msg.
2.)	Remove all instances of busy_lock. busy_lock was renamed to
	cam_lock and used to lock out simultaneous uses of the camera
	and handle disconnects. We may want to add back a different lock
	to handle smp type stuff.
3.)	Separate read_frame and vicam_decode_color. This should move us
	along toward asynchronous urbs.

This patch does not address the locking of the camera that is still
needed by the proc interface.

2002/11/25 16:39:49-08:00 greg
[PATCH] USB: added Palm Tungsten W support.

Thanks to Ralf Dietrich <ralle@envicon.de> for the information.

2002/11/25 10:53:08-08:00 greg
[PATCH] driver core: fix compiler warning if CONFIG_HOTPLUG is not defined

fixes bug #26 <http://bugme.osdl.org/show_bug.cgi?id=26>

2002/11/23 23:52:53-08:00 baldrick
[PATCH] usbdevfs: finalize urbs on interface release

Description: When an urb has been submitted via usbdevfs, and is still
pending when the interface it was submitted to is released, force the
urb to be completed.  This is the correct behaviour.  It fixes an oops on
system shutdown when using the user space driver for the speedtouch
modem.

2002/11/23 23:47:11-08:00 ganesh
[PATCH] uninitialized spinlock in ipaq.c

this fixes an uninitialzed spinlock in ipaq.c. the driver should work
on smp machines now.

2002/11/23 23:37:01-08:00 randy.dunlap
[PATCH] USB: use time_before() to compare times

2002/11/23 23:34:26-08:00 david-b
[PATCH] ohci uses less slab memory

When chasing down some of those 'bad entry' diagnostics, I once
got suspicious that the problem was slab corruption coming from
the way the td hashtable code worked.   So I put together this
patch, eliminating some kmallocation, and the next times I ran
that test, the oops went away and it worked like a charm.  Hmm.

This patch is good because it shrinks memory and code, and gets
rid of some could-fail allocations, so I figured I'd send it on
(low priority) even if I don't think it fixes the root problem.

2002/11/23 23:33:53-08:00 david-b
[PATCH] remove CONFIG_USB_LONG_TIMEOUT

Basically, no point in having short and long timeout options
where both are _shorter_ than the timeout from the USB spec.

2002/11/23 23:33:37-08:00 david-b
[PATCH] drivers/base/hotplug.c, fix $DEVPATH

Hotplug agents couldn't use /sys/$DEVPATH after /sys/root
morphed into /sys/devices ... now they can do it again.

BKrev: 3de51c00552mUbEsFOCyIRijefgApg