[SERIAL] Prevent PNPBIOS re-registering already detected ports
During initialisation of 8250 serial, we scan a list of ISA ports and
register any ports.  We then perform PNPBIOS scanning, which re-registers
ttyS0.  Unfortunately, if devfs is enabled, devfs reports an error because
we try to create two tts/0 entries.

Therefore, when adding a new port we check that it has not been detected
before attempting to probe the port and register it with devfs (via the
tty layer.)

BKrev: 3de9dd935DtWOxjs4nIzXl1SmcbDaQ