Merge kroah.com:/home/greg/linux/BK/bleeding_edge-2.5 into kroah.com:/home/greg/linux/BK/pnp-2.5
2002/12/06 10:08:34-06:00 ambx1
[PATCH] PnP gameport driver update

This trivial patch updates the gameport driver to the new id scheme.

2002/12/05 23:31:41-06:00 ambx1
[PATCH] PnP bugfix

I forgot the errno.h.  Without this patch, things may not compile when
pnp support or pnp card support is disabled.

Hope you had a good trip.  I noticed a small mistake in pnp 0.93 and I have
a fix for it.

2002/12/05 23:31:21-06:00 ambx1
[PATCH] Linux PnP Support V0.93 - 2.5.50

Attached is a patch, that updates the 2.5.50 to the latest pnp
version.  It includes all 9 of the previously submitted patches.

Highlights are as follows:
-PnP BIOS fixes
-Several new macros
-PnP Card Services
-Various bug fixes
-more drivers converted to the new APIs

BKrev: 3df0cf6d4FVUKndhbfxjL7pksw5PGA