Merge bk://linuxusb.bkbits.net/pnp-2.5 into home.transmeta.com:/home/torvalds/v2.5/linux
2002/12/06 08:33:23-08:00 markh
[PATCH] aacraid 2.5 update

This contains three changes.  The first removes some #defines that don't
seem to be needed anymore.  The second fixes a problem when CONFIG_LBD
is set and sector_t is u64.  The third fixes some compile warnings
setting a char * in the scsi_pointer struct to a dma_addr_t.  I changed
the usage from the ptr element to the dma_handle element.

It compiles without warnings and I have run some tests on aacraid devices
on the osdl lab machines with the CONFIG_LBD option on and off.

BKrev: 3df0d5d6I-8gBrsZ3gKzVX8a_H5Irw