[PATCH] remove error message on illegal ioctl
This error message is uber annoying and needs to go.  Non-root can flood
the console with this junk on invalid SCSI CD-ROM ioctl(), and that is
exactly what gnome-cd does.

An illegal ioctl() returns an error to the program.  That is sufficient
- we do not need KERN_ERROR warnings all over the place.  Especially
when any user can cause them at any rate.

BKrev: 3dfaa79coIXmKv2UfGP9no5jV6u_rw