[PATCH] printks in drivers/scsi/hosts.c missing return
Trivial but annoying: two printk() calls in drivers/scsi/hosts.c are
missing '\n'

Also, for some reason I have not yet investigated, shost_tp->name is
NULL here.  This should not be, eh?  If it can be, we should do
something to tidy up the printing of it.

BKrev: 3dfaa7a1ovohZ5Z4-P2amSbWhUwjBg