[PATCH] Module Parameter Core Patch
This patch is a rewrite of the insmod and boot parameter handling,
to unify them.

The new format is fairly simple: built on top of __module_param_call there
are several helpers, eg "module_param(foo, int, 000)".  The final argument
is the permissions bits, for exposing parameters in sysfs (if
non-zero) at a later stage.

BKrev: 3dfce257LoE-uARwM9aAb4Qwsf5MDw