[PATCH] deprecated function attribute
This patch adds support for usage of the attribute as "deprecated" and
is backward-compatible.  Usage is:

	int deprecated foo(void)

etc..

If we mark a function as deprecated, then each use of the function emits
a warning like:

	foo.c:12: warning: `baz' is deprecated (declared at bar.c:60)

Which is very informative, giving both the location of each usage and
where the little bastard is declared.

BKrev: 3e0e758cLbYbDy-khhDzwRrgNbH7lA