[PATCH] Modules 2/3: Use sh_addr instead of sh_offset
From: Richard Henderson <rth@twiddle.net>

Richard points out that we should be using sh_addr to hold the address,

The original patch used to overload sh_offset to a pointer to the
location of the section.  This uses sh_addr, which is more correct
and less surprising.

BKrev: 3e13abc8kWMmUim_nZFqVtS55XS4sA