[PATCH] Introduce TIF_IRET and use it to disable sysexit
This patch introduces a new flag called TIF_IRET and uses it in place
of TIF_SIGPENDING when that flag is used to force return via iret.

This avoids the overhead of calling do_signal and makes the code
easier to understand.

BKrev: 3e18f633g_BhuEBeJSg0h4Dk-WvU_A