[PATCH] knfsd: Add 'threads' file to nfsd filesystem to allow changing number of threads.
This uses the read-without-write style transaction files in nfsctl.
We can write a number of threads, and then read back the number of
threads that resulted, or we can just open and read in which case
we read back the number of threads without changing it.

BKrev: 3e18fb63hLraFNerurnN8BSHitt7nw