[PATCH] add skb_pad/skb_padto functionality
This is the stuff in 2.4 which DaveM has been over so I hope is ok for 2.5.
It provides a simple API for driver writers with hardware that doesn't
do packet padding to say 'make sure n bytes beyond the data is all clear'
for DMA etc.

BKrev: 3e1f99dcFYrJ4qLu1PY7pj90Q9pXTQ